% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heleramcar5GUI-package.R
\docType{package}
\name{heleramcar5GUI-package}
\alias{heleramcar5GUI}
\alias{heleramcar5GUI-package}
\title{heleramcar5GUI: GUI for heleramcar5 R package}
\description{
GUI for heleramcar5 R package using Shiny.
}
\author{
\strong{Maintainer}: Helena Llorens \email{hllor282@student.liu.se}

Authors:
\itemize{
  \item Carmen Rodríguez \email{carro162@student.liu.se}
  \item Ramón Amézquita \email{ramam120@student.liu.se}
}

}
\keyword{internal}
