% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Define Shiny server logic.}
\usage{
server(input, output)
}
\arguments{
\item{input}{Shiny input environment}

\item{output}{Shiny output environment}
}
\description{
Define Shiny server logic.
}
